remove some files and resolve the PR comments
