+++ src/apps/frontend/pages/tasks/task-card.tsx
@@ -1,9 +1,9 @@
 import React, { PropsWithChildren } from 'react';
 
-const TaskItemCard: React.FC<PropsWithChildren> = ({ children }) => (
+const TaskCard: React.FC<PropsWithChildren> = ({ children }) => (
   <div className="relative mt-5 rounded-sm border border-stroke bg-white p-6 shadow-default">
     {children}
   </div>
 );
 
-export default TaskItemCard;
+export default TaskCard;
