+++ src/apps/frontend/pages/tasks/task-header.tsx
@@ -7,7 +7,7 @@ import {
 import { AsyncError } from '../../types';
 import { TaskOperationType } from '../../types/task';
 
-import TaskItemCard from './task-item-card';
+import TaskCard from './task-card';
 import TaskModal from './task-modal';
 import useTaskForm from './tasks-form.hook';
 
@@ -29,27 +29,27 @@ const TaskHeader: React.FC<TaskHeaderProps> = ({ onError }) => {
   });
 
   return (
-    <TaskItemCard>
-      <Flex alignItems='center' justifyContent='between'>
-        <FlexItem >
+    <TaskCard>
+      <Flex alignItems="center" justifyContent="between">
+        <FlexItem>
           <HeadingLarge>Tasks</HeadingLarge>
         </FlexItem>
-        <FlexItem >
+        <FlexItem>
           <Button onClick={() => setIsModalOpen(!isModalOpen)}>
             <img src="assets/svg/plus-icon.svg" alt="Plus Icon" />
             Add task
           </Button>
-        <TaskModal
-          formik={addTaskFormik}
-          setFieldValue={setFormikFieldValue}
-          isModalOpen={isModalOpen}
-          setIsModalOpen={setIsModalOpen}
-          btnText={'Add Task'}
-          taskOperationType={TaskOperationType.ADD}
-        />
+          <TaskModal
+            formik={addTaskFormik}
+            setFieldValue={setFormikFieldValue}
+            isModalOpen={isModalOpen}
+            setIsModalOpen={setIsModalOpen}
+            btnText={'Add Task'}
+            taskOperationType={TaskOperationType.ADD}
+          />
         </FlexItem>
       </Flex>
-    </TaskItemCard>
+    </TaskCard>
   );
 };
